+++ src/locales/en.json
@@ -2862,11 +2862,11 @@
   "scope.text_230": "{0} Information",
   "scope.text_231": "Currently logged in {0}",
   "scope.text_232": "Manage {0}",
-  "scope.text_233": "Contact Information",
+  "scope.text_233": "Notification Message Settings",
   "scope.text_234": "Modify",
   "scope.text_235": "Mobile:",
   "scope.text_236": "Email:",
-  "scope.text_237": "Enable Status:",
+  "scope.text_237": "Notification Message Enable Status:",
   "scope.text_238": "Preferences",
   "scope.text_239": "Show System Resources:",
   "scope.text_240": "ON",
@@ -3576,7 +3576,7 @@
   "scope.text_954": "One Year",
   "scope.text_955": "The number cannot be 0",
   "scope.text_956": "OpenID Connect connection information:",
-  "scope.text_957": "Not verified, please click the icon to verify",
+  "scope.text_957": "Not verified",
   "scope.text_958": "Backup：{0}",
   "scope.text_959": "IAM & Security",
   "scope.text_960": "Security Inspection",
@@ -4404,6 +4404,7 @@
   },
   "common_730": "View Tag",
   "common_731": "No Channel",
+  "common.verify": "Verify",
   "common.search": "Please enter search content",
   "common_732": "No more than 20 tags",
   "common_733": "Total: {0}",
