+++ src/locales/zh-CN.json
@@ -2875,11 +2875,11 @@
   "scope.text_230": "{0}信息",
   "scope.text_231": "当前登录{0}",
   "scope.text_232": "管理{0}",
-  "scope.text_233": "联系方式",
+  "scope.text_233": "消息接收设置",
   "scope.text_234": "修改",
-  "scope.text_235": "手机：",
+  "scope.text_235": "手机号码：",
   "scope.text_236": "邮箱：",
-  "scope.text_237": "启用状态：",
+  "scope.text_237": "接收消息通知：",
   "scope.text_238": "偏好设置",
   "scope.text_239": "是否显示系统资源：",
   "scope.text_240": "开",
@@ -3589,7 +3589,7 @@
   "scope.text_954": "一年",
   "scope.text_955": "金额不能为0",
   "scope.text_956": "OpenID Connect连接信息：",
-  "scope.text_957": "未验证，请点击图标验证",
+  "scope.text_957": "未验证",
   "scope.text_958": "备：{0}",
   "scope.text_959": "认证与安全",
   "scope.text_960": "安全检查",
@@ -4417,6 +4417,7 @@
   },
   "common_730": "查看标签",
   "common_731": "暂无通知渠道",
+  "common.verify": "验证",
   "common.search": "请输入搜索内容",
   "common_732": "标签不可超过20个",
   "common_733": "总计: {0}",
